backend/requirements.txt
flask
requests
gunicorn
python-dotenv
Flask-SQLAlchemy
pymysql
Pillow
